requirements.txt
keyboard==0.13.5
numpy==2.1.3
opencv_python==4.9.0.80
Pillow==11.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
